+++ NEWS.md
@@ -42,7 +42,7 @@
     XEX executables is supported at the moment.
   * Xbox360_XEX: Xbox 360 executables.
   * BRSTM: Nintendo Wii BRSTM audio files.
-  * BCSTM: Nintendo 3DS BCSTM audio files.
+  * BCSTM: Nintendo 3DS BCSTM and BCWAV audio files.
     * Also supports Nintendo Wii U BFSTM audio files.
 
 * New audio parsers: